// workplace-fe/src/app/pages/storage/model/deck.ts
export class Deck {
  id: string | undefined;
  adminId: string | undefined;
  workplaceId: string | undefined;
  name: string | undefined;
  customName: string | undefined;
  description: string | undefined;
  customDescription: string | undefined;
  cardBack: string | undefined;
  cardBackUrl: string | undefined;
  cardBackExtension: string | undefined;
  cardsCount: string | undefined;
  available: boolean | undefined;
  deckInCollection: boolean | undefined;
}